Work on sem tokens
